Update UI (#34)
* scoring method ui updates

* update server format

* update test

* update requirements freeze

* remove circular import with shared utils package

* fix github unit test running

* update test import path

* update github workflow

* update test reqs
