+++ arthur_bench/server/js/packages/arthur-redux/slices/testSuites/types.ts
@@ -1,7 +1,7 @@
 export type TTestSuite = {
     id: string;
     name: string;
-    scoring_method: EMethodType;
+    scoring_method: TScoringMethod;
     last_run_time: string;
     updated_at: string;
     created_at: string;
@@ -39,7 +39,7 @@ export type DetailedTestSuite = {
     id: string;
     description: string;
     name: string;
-    scoring_method: EMethodType;
+    scoring_method: TScoringMethod;
     last_run_time: string;
     num_runs: number;
     test_cases: TestSuiteCase[];
@@ -102,6 +102,11 @@ export type TTestSuitesState = {
     currentTestRun: TTestRun;
     pagination: TPagination;
 };
+
+export type TScoringMethod = {
+    name: EMethodType;
+    type: string
+}
 
 export enum EMethodType {
     BERT = 'bertscore',
