+++ arthur_bench/version.py
@@ -6,4 +6,4 @@
 # MAJOR version when you make incompatible API changes,
 # MINOR version when you add functionality in a backwards compatible manner, and
 # PATCH version when you make backwards compatible bug fixes.
-__version__ = "0.1.1"
+__version__ = "0.1.2"
