+++ pyproject.toml
@@ -57,7 +57,7 @@ bench = "arthur_bench.server.run_server:run"
 url = 'http://arthur.ai'
 
 [tool.setuptools.packages.find]
-exclude = ["docs*", "test*", "examples*"]
+exclude = ["docs*", "tests*", "examples*"]
 
 [tool.setuptools.dynamic]
 version = {attr = "arthur_bench.__version__"}
