+++ requirements.txt
@@ -1,13 +1,10 @@
 aiohttp==3.8.5
 aiosignal==1.3.1
-amplitude-analytics==1.1.1
-anyio==3.7.1
-async-timeout==4.0.2
+async-timeout==4.0.3
 attrs==23.1.0
 bert-score==0.3.13
 certifi==2023.7.22
 charset-normalizer==3.2.0
-click==8.1.6
 contourpy==1.1.0
 cycler==0.11.0
 dataclasses-json==0.5.14
@@ -20,14 +17,13 @@ fonttools==4.42.0
 frozenlist==1.4.0
 fsspec==2023.6.0
 greenlet==2.0.2
-h11==0.14.0
 huggingface-hub==0.16.4
 idna==3.4
 importlib-resources==6.0.1
 Jinja2==3.1.2
 kiwisolver==1.4.4
-langchain==0.0.258
-langsmith==0.0.19
+langchain==0.0.264
+langsmith==0.0.22
 MarkupSafe==2.1.3
 marshmallow==3.20.1
 matplotlib==3.7.2
@@ -45,29 +41,27 @@ Pillow==10.0.0
 pydantic==1.10.12
 PyJWT==2.8.0
 pyparsing==3.0.9
+pyphen==0.14.0
 python-dateutil==2.8.2
 pytz==2023.3
 PyYAML==6.0.1
-regex==2023.6.3
+regex==2023.8.8
 requests==2.28.2
 requests-toolbelt==1.0.0
 sentencepiece==0.1.99
 six==1.16.0
-sniffio==1.3.0
 SQLAlchemy==2.0.19
-starlette==0.27.0
 sympy==1.12
-tenacity==8.2.2
+tenacity==8.2.3
 textstat==0.7.3
 tiktoken==0.4.0
 tokenizers==0.13.3
 torch==2.0.1
-tqdm==4.65.1
+tqdm==4.65.2
 transformers==4.29.2
 typing-inspect==0.9.0
 typing_extensions==4.7.1
 tzdata==2023.3
 urllib3==1.26.16
-uvicorn==0.22.0
 yarl==1.9.2
 zipp==3.16.2
