--- test/requirements.txt
@@ -1,3 +1,0 @@
-fastapi==0.99.1
-httpx==0.24.1
-pytest==7.3.1
