+++ tests/fixtures/mock_suite_mixed_null_refs.json
