+++ tests/fixtures/mock_suite_with_typed_score.json
