+++ tests/fixtures/mock_suite_without_optional.json
