+++ tests/requirements.txt
@@ -1,0 +1,5 @@
+amplitude-analytics==1.1.1
+fastapi==0.99.1
+httpx==0.24.1
+pytest==7.3.1
+uvicorn~=0.22.0
