+++ tests/test_testrun.py
@@ -1,5 +1,6 @@
 import pytest
-from fixtures.mock_requests import MOCK_RUN, MOCK_RUN_JSON
+from tests.fixtures.mock_requests import MOCK_RUN, MOCK_RUN_JSON
+
 
 def test_run_serialization():
     assert MOCK_RUN.json() == MOCK_RUN_JSON
