+++ tests/test_utils.py
@@ -1,7 +1,7 @@
 import pytest
 import pandas as pd
-from helpers import FIXTURE_FILE_DIR
-from fixtures.mock_responses import (
+from tests.helpers import FIXTURE_FILE_DIR
+from tests.fixtures.mock_responses import (
     MOCK_SUITE_RESPONSE,
     MOCK_SUITE_JSON,
     MOCK_SUITE_WITH_NULL,
@@ -9,7 +9,7 @@ from fixtures.mock_responses import (
     MOCK_SUITE_WITH_SCORING_CONFIG,
     MOCK_SUITE_WITH_SCORING_JSON,
 )
-from arthur_bench.run.utils import (
+from arthur_bench.utils.loaders import (
     load_suite_from_json,
     load_suite_from_dataframe,
     load_suite_from_list,
